feat: modularization of functions
