+++ html/1.0.0/js/script.js
@@ -11,6 +11,11 @@ import {
 	validateForm, resetErrorField
 } from './modules/formsValidators.js'
 
+import {
+	getEncodeURLEmail, loadEmail
+} from './modules/decodeEmail.js'
+
+
 document.addEventListener('DOMContentLoaded', () => {
 
 
@@ -23,42 +28,8 @@ document.addEventListener('DOMContentLoaded', () => {
 	const utm_medium = (document.getElementById("utm_medium").value != '' ? document.getElementById("utm_medium").value : '');
 	const dialCodeContainer = document.querySelector('.iti__selected-dial-code');
 
-	checkEncodeURLEmail()
-
-	function getURLParam(searchedParam) {
-		const queryString = window.location.search;
-		const urlParams = new URLSearchParams(queryString);
-		return urlParams.get(searchedParam);
-
-	}
-
-	// Function to decode email
-	function fromHex(hex) {
-		var str = '';
-		for (var i = 0; i < hex.length; i += 2) {
-			var v = parseInt(hex.substr(i, 2), 16);
-			if (v) str += String.fromCharCode(v);
-		}
-		return str;
-	}
-
-
-
-	function checkEncodeURLEmail() {
-		const urlParamHash = getURLParam('dplrid')
-
-		if (urlParamHash) {
-			let urlEmailDecode = fromHex(urlParamHash);
-			loadEmail(urlEmailDecode);
-		}
-	}
-
-
-	function loadEmail(urlEmailDecode) {
-		document.querySelectorAll('form').forEach( form => {
-			form.querySelector('input[name="email"]').value = urlEmailDecode
-		})
-	}
+	const urlEmail = getEncodeURLEmail();
+	loadEmail(urlEmail);
 
 
 	earlyForms.forEach(form => { form.addEventListener('submit', sendData) });
